Refactor: Extract new crate binstalk-manifests
