+++ crates/binstalk-manifests/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "binstalk-manifests"
+description = "The binstall toolkit for manipulating with manifest"
+repository = "https://github.com/cargo-bins/cargo-binstall"
+documentation = "https://docs.rs/binstalk-manifests"
+version = "0.1.0"
+rust-version = "1.61.0"
+authors = ["ryan <ryan@kurte.nz>"]
+edition = "2021"
+license = "GPL-3.0"
+
+[dependencies]
+compact_str = { version = "0.6.0", features = ["serde"] }
+fs-lock = { version = "0.1.0", path = "../fs-lock" }
+home = "0.5.4"
+miette = "5.4.1"
+once_cell = "1.16.0"
+semver = { version = "1.0.14", features = ["serde"] }
+serde = { version = "1.0.147", features = ["derive"] }
+serde-tuple-vec-map = "1.0.1"
+serde_json = "1.0.87"
+strum = "0.24.1"
+strum_macros = "0.24.3"
+thiserror = "1.0.37"
+toml_edit = { version = "0.15.0", features = ["easy"] }
+url = { version = "2.3.1", features = ["serde"] }
+
+[dev-dependencies]
+detect-targets = { version = "0.1.2", path = "../detect-targets" }
+tempfile = "3.3.0"
