+++ crates/binstalk-manifests/src/cargo_crates_v1/crate_version_source.rs
@@ -8,8 +8,8 @@ use thiserror::Error;
 use url::Url;
 
 use crate::{
-    helpers::statics::cratesio_url,
-    manifests::crate_info::{CrateInfo, CrateSource, SourceType},
+    crate_info::{CrateInfo, CrateSource, SourceType},
+    helpers::cratesio_url,
 };
 
 #[derive(Clone, Debug, Ord, PartialOrd, Eq, PartialEq)]
