+++ crates/binstalk-manifests/src/cargo_toml_binstall.rs
