+++ crates/binstalk-manifests/src/cargo_toml_binstall/package_formats.rs
