+++ crates/binstalk-manifests/src/crate_info.rs
@@ -7,7 +7,7 @@ use semver::Version;
 use serde::{Deserialize, Serialize};
 use url::Url;
 
-use crate::helpers::statics::cratesio_url;
+use crate::helpers::cratesio_url;
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct CrateInfo {
