+++ crates/binstalk/src/fs.rs
@@ -3,20 +3,6 @@ use std::{fs, io, path::Path};
 use log::debug;
 use tempfile::NamedTempFile;
 
-/// Returned file is readable and writable.
-pub fn create_if_not_exist(path: impl AsRef<Path>) -> io::Result<fs::File> {
-    let path = path.as_ref();
-
-    let mut options = fs::File::options();
-    options.read(true).write(true);
-
-    options
-        .clone()
-        .create_new(true)
-        .open(path)
-        .or_else(|_| options.open(path))
-}
-
 /// Atomically install a file.
 ///
 /// This is a blocking function, must be called in `block_in_place` mode.
