+++ crates/binstalk/src/helpers.rs
@@ -2,5 +2,4 @@ pub mod download;
 pub mod jobserver_client;
 pub mod remote;
 pub mod signal;
-pub mod statics;
 pub mod tasks;
